refactor: split receipt validation helpers
